split up glossary component
